Move HeaderBackgroundImage component to components folder
