+++ src/components/index.js
@@ -1,4 +1,5 @@
 import ScrollableTabBar from './ScrollableTabBar/ScrollableTabBar';
 import ScrollableTabView from './ScrollableTabView/ScrollableTabView';
+import HeaderBackgroundImage from './HeaderBackgroundImage/HeaderBackgroundImage';
 
-export { ScrollableTabBar, ScrollableTabView };
+export { ScrollableTabBar, ScrollableTabView, HeaderBackgroundImage };
